server/requirements.txt
alembic==1.7.5
aniso8601==9.0.1
attrs==21.2.0
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-JWT-Extended==4.3.1
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.2
jsonschema==4.1.2
Mako==1.1.6
MarkupSafe==2.0.1
PyJWT==2.3.0
pyrsistent==0.18.0
python-decouple==3.5
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.26
Werkzeug==2.0.2
